added authorization to aliasController and tlspolicyController

+++ src/controllers/aliasController.js
@@ -1,8 +1,15 @@
 import { isEmpty } from "lodash";
 import Alias from "../model/alias";
 
+import { isAdmin } from "../helpers/authorizationHelper";
+
 class AliasController {
   async index(req, res) {
+    if (!isAdmin(req.user.email))
+      return res
+        .status(403)
+        .json({ message: "only admins are allowed to view all aliases" });
+
     const aliases = await Alias.getAliases();
     res.json({ aliases });
   }
@@ -15,6 +22,13 @@ class AliasController {
 
     const alias = (await Alias.getAlias(id))[0];
     if (alias) {
+      const [, domain] = req.user.email.split("@");
+      if (!isAdmin(req.user.email) && alias.source_domain !== domain)
+        return res.status(403).json({
+          message:
+            "only admins and the user who this alias belongs to are allowed to view this alias"
+        });
+
       res.json({ alias });
     } else {
       res.status(404).json({ message: "alias not found" });
@@ -37,6 +51,13 @@ class AliasController {
       !destination_domain
     )
       return res.status(400).json({ message: "parameters are missing" });
+
+    const [, domain] = req.user.email.split("@");
+    if (!isAdmin(req.user.email) && source_domain !== domain)
+      return res.status(403).json({
+        message:
+          "only admins can create aliases for other domains than the one he belongs to"
+      });
 
     const id = (await Alias.createAlias({
       source_username,
@@ -69,6 +90,13 @@ class AliasController {
     if (!id)
       return res.status(400).json({ message: "parameter id is missing" });
 
+    const [, domain] = req.user.email.split("@");
+    if (!isAdmin(req.user.email) && source_domain !== domain)
+      return res.status(403).json({
+        message:
+          "only admins can update aliases for other domains than the one he belongs to"
+      });
+
     await Alias.updateAlias(
       {
         source_username,
@@ -89,9 +117,18 @@ class AliasController {
   async delete(req, res) {
     const { id } = req.params;
 
-    const rowsDeleted = await Alias.deleteAlias(id);
-    if (rowsDeleted === 0)
-      return res.status(404).json({ message: "alias not found" });
+    const alias = (await Alias.getAlias(id))[0];
+    if (!alias) return res.status(404).json({ message: "alias not found" });
+
+    const [, domain] = req.user.email.split("@");
+    if (!isAdmin(req.user.email) && alias.source_domain !== domain)
+      return res.status(403).json({
+        message:
+          "only admins can delete aliases for other domains than the one he belongs to"
+      });
+
+    await Alias.deleteAlias(id);
+
     return res.status(204).end();
   }
 }
